Add patch to fix shared lib
Hsakmt is not currently designed to work as a shared lib. I spoke to
upstream about this and it appears that they want to merge it into
rocm-runtime soon (possibly as early as 6.1).

To avoid turning this into a static lib, I just patched the symbol list
to work with shared linking, see:
https://github.com/ROCm/ROCT-Thunk-Interface/pull/90

But in the long term we will retire this package when upstream become
abandoned due to the merge